pyproject.toml
[tool.poetry]
name = "erp-tender"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "erp_tender"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
django-import-export = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
